+++ test/util/util.ts
@@ -19,7 +19,13 @@ import { createInMemoryFtKeyStore } from "/ft4/authentication/ft/key-stores/in-m
 import { createAuthenticator } from "/ft4/authentication";
 import { transactionBuilder } from "/ft4/utils/transaction-builder";
 import { addAuthDescriptor } from "/ft4/accounts/account-operations";
-import { createAuthDataService, createConnection } from "/ft4/ft-session";
+import {
+  createAuthDataService,
+  createConnection,
+  createKeyStoreInteractor,
+} from "/ft4/ft-session";
+import { BufferId } from "/cryptoUtils";
+import { Connection } from "/ft4";
 
 function generateNumber(max = 10000): number {
   return Math.round(Math.random() * max);
@@ -80,7 +86,7 @@ export {
 
 export function createTestAuthDescriptor(
   flags: string[] = [],
-  rules?: AuthDescriptorRule
+  rules?: AuthDescriptorRule,
 ): {
   keyPair: KeyPair;
   authDescriptor: AuthDescriptor;
@@ -94,18 +100,18 @@ export function createTestAuthDescriptor(
 
 export function createTestMultisigAuthDescriptor(
   requiredSignatures: number,
-  flags: string[] = []
+  flags: string[] = [],
 ): {
   keyPairs: KeyPair[];
   authDescriptor: AuthDescriptor;
 } {
   const keyPairs = Array.from({ length: requiredSignatures }, () =>
-    encryption.makeKeyPair()
+    encryption.makeKeyPair(),
   );
   const descriptor = authDescriptor.create.multiSig.withArgs(
     flags,
     requiredSignatures,
-    keyPairs.map((kp) => kp.pubKey)
+    keyPairs.map((kp) => kp.pubKey),
   ).andNoRules;
 
   return { keyPairs, authDescriptor: descriptor };
@@ -121,21 +127,21 @@ export async function addAuthDescriptorTo(
   newUser: {
     signatureProvider: SignatureProvider;
     authDescriptor: AuthDescriptor;
-  }
+  },
 ) {
   const keyHandlerUser1 = createInMemoryFtKeyStore(
-    user.signatureProvider
+    user.signatureProvider,
   ).createKeyHandler(user.authDescriptor);
 
   const keyHandlerUser2 = createInMemoryFtKeyStore(
-    newUser.signatureProvider
+    newUser.signatureProvider,
   ).createKeyHandler(newUser.authDescriptor);
 
   const authDataService = createAuthDataService(createConnection(client));
   const authenticator = createAuthenticator(
     accountId,
     [keyHandlerUser1],
-    authDataService
+    authDataService,
   );
 
   const tx = await transactionBuilder(authenticator, client)
@@ -147,14 +153,27 @@ export async function addAuthDescriptorTo(
 
 export async function createAccount(
   client: IClient,
-  descriptor: AuthDescriptor
+  descriptor: AuthDescriptor,
 ) {
   const ad = authDescriptor.toGtv(descriptor);
   await client.signAndSendUniqueTransaction(
     op("register_account_test", [ad[1], ad[2], ad[3]]),
-    adminUser().signatureProvider
+    adminUser().signatureProvider,
   );
   return descriptor.id;
+}
+
+export async function getSessionForAccount(
+  connection: Connection,
+  accountId: BufferId,
+  signer: SignatureProvider | KeyPair,
+) {
+  const { getSession } = createKeyStoreInteractor(
+    connection.client,
+    createInMemoryFtKeyStore(signer),
+  );
+
+  return await getSession(accountId);
 }
 
 export function rellError(message: string) {